// src/interfaces/IWechatMessage.ts
export interface IWechatMessage {
  ToUserName: string;
  FromUserName: string;
  CreateTime: string;
  MsgType: string;
  MsgId: string;
  Content: string;
  PicUrl: string;
  MediaId: string;
  Format: string;
  Recognition: string;
  ThumbMediaId: string;
  Title: string;
  Description: string;
  Url: string;
  Location_X: string;
  Location_Y: string;
  Scale: string;
  Label: string;
}